Fixed GiteeConnectionConfigTest with repo's implementation, added tests, updated dependencies to pom.xml
